Added a test to show the issue in PLFM-112 "the default response encoding for the service seems to vary between XML and JSON"
We would need to refactor a few of our model objects to get JAXB to bind to them correctly.

I think the longer term fix is to just disable XML encoding to avoid confusion.  If/when we decide we need XML, we should ensure that it is supported for all model objects.