Okay, this actually works.  It worked just fine with the SWF service and ran all the way through to completion for the first four of of the jobs in Brig's test spreadsheet.  I cut it off at four to stay under the 32k limit on output from an activity.
Since a workflow can only kick off at most 100 activities per decision, using child workflows is still the right approach for this, but the way I have it now it returns too much data.  I need to either (1) move the parsing of the spreadsheet into the decider or (2) figure out the right way to kick off workflows from an activity.