Fix for PLFM-1020 'Prefix synapse id's with "SYN" in all service communications'
All entity ids leaving the DBO/JDO layer should be prefixed with "syn".  This includes not only entity ids, but also parentIds, targetIds in references, and child ids retrieved via create/update/get/query.  This does not include etags or the soon-to-be-deprecated eula ids.

Note that there is one special case in the repository service S3 keys - the syn prefix is removed because we don't want to move all that data around in S3 _again_ just to have it match this new prefix.